+++ src/component/view/ViewManager.js
@@ -10,21 +10,21 @@ var ViewManager = new function (){
      * @param  {View} targetView 目标视图
      */
 	this.replace = function(newView,targetView){
-		var targetV = targetView.elements[0],
-			newV = newView.elements[0],
-			p = targetView.elements[0].parentNode;
+		var targetV = targetView.__nodes[0],
+			newV = newView.__nodes[0],
+			p = targetView.__nodes[0].parentNode;
 		var fragment = newV;
-		if(newView.elements.length > 1){
+		if(newView.__nodes.length > 1){
 			fragment = document.createDocumentFragment();
-			for(var i=0;i<newView.elements.length;i++){
-				fragment.appendChild(newView.elements[i]);
+			for(var i=0;i<newView.__nodes.length;i++){
+				fragment.appendChild(newView.__nodes[i]);
 			}
 		}
-		if(targetView.elements.length > 1){
+		if(targetView.__nodes.length > 1){
 			p.insertBefore(fragment,targetV);
 			
-			for(var i=targetView.elements.length;i--;){
-				p.removeChild(targetView.elements[i]);
+			for(var i=targetView.__nodes.length;i--;){
+				p.removeChild(targetView.__nodes[i]);
 			}
 		}else{
 			p.replaceChild(fragment,targetV);
@@ -37,14 +37,14 @@ var ViewManager = new function (){
 	 * @param  {View} targetView 目标视图
 	 */
 	this.insertBefore = function(newView,targetView){
-		var targetV = targetView.elements[0],
-			newV = newView.elements[0],
-			p = targetView.elements[0].parentNode;
+		var targetV = targetView.__nodes[0],
+			newV = newView.__nodes[0],
+			p = targetView.__nodes[0].parentNode;
 		var fragment = newV;
-		if(newView.elements.length > 1){
+		if(newView.__nodes.length > 1){
 			fragment = document.createDocumentFragment();
-			for(var i=0;i<newView.elements.length;i++){
-				fragment.appendChild(newView.elements[i]);
+			for(var i=0;i<newView.__nodes.length;i++){
+				fragment.appendChild(newView.__nodes[i]);
 			}
 		}
 		if(p)
@@ -57,24 +57,62 @@ var ViewManager = new function (){
 	 * @param  {View} targetView 目标视图
 	 */
 	this.insertAfter = function(newView,targetView){
-		var targetV = targetView.elements[0],
-			newV = newView.elements[0],
-			p = targetView.elements[0].parentNode;
+		var targetV = targetView.__nodes[0],
+			newV = newView.__nodes[0],
+			p = targetView.__nodes[0].parentNode;
 		var last = p.lastChild;
 		var fragment = newV;
-		if(newView.elements.length > 1){
+		if(newView.__nodes.length > 1){
 			fragment = document.createDocumentFragment();
-			for(var i=0;i<newView.elements.length;i++){
-				fragment.appendChild(newView.elements[i]);
+			for(var i=0;i<newView.__nodes.length;i++){
+				fragment.appendChild(newView.__nodes[i]);
 			}
 		}
 
 		if(last == targetV){
 			p.appendChild(fragment);
 		}else{
-			var next = targetView.elements[targetView.elements.length-1].nextSibling;
+			var next = targetView.__nodes[targetView.__nodes.length-1].nextSibling;
 			p.insertBefore(fragment,next);
 		}
+	}
+
+	/**
+	 * 在指定视图内插入视图，并插入到最后位置
+	 * @param  {View} newView 新视图
+	 * @param  {View} targetView 目标视图
+	 */
+	this.append = function(newView,targetView){
+		var newV = newView.__nodes[0],
+			p = targetView.__nodes[0];
+		var fragment = newV;
+		if(newView.__nodes.length > 1){
+			fragment = document.createDocumentFragment();
+			for(var i=0;i<newView.__nodes.length;i++){
+				fragment.appendChild(newView.__nodes[i]);
+			}
+		}
+
+		p.appendChild(fragment);
+	}
+
+	/**
+	 * 在指定视图内插入视图，并插入到最前位置
+	 * @param  {View} newView 新视图
+	 * @param  {View} targetView 目标视图
+	 */
+	this.prepend = function(newView,targetView){
+		var newV = newView.__nodes[0],
+			p = targetView.__nodes[0];
+		var fragment = newV;
+		if(newView.__nodes.length > 1){
+			fragment = document.createDocumentFragment();
+			for(var i=0;i<newView.__nodes.length;i++){
+				fragment.appendChild(newView.__nodes[i]);
+			}
+		}
+
+		p.insertBefore(fragment,p.firstChild);
 	}
 
 	/**
@@ -84,6 +122,6 @@ var ViewManager = new function (){
 	 * @return {View} 新视图
 	 */
 	this.createPlaceholder = function(content){
-		return new View([document.createComment(content)]);
+		return new View(null,null,[document.createComment(content)]);
 	}
 }